+++ examples/data-sources/pingfederate_session_authentication_session_policies_global/data-source.tf
@@ -1,0 +1,2 @@
+data "pingfederate_session_authentication_session_policies_global" "mySessionAuthenticationSessionPoliciesGlobalExample" {
+}