+++ internal/provider/provider.go
@@ -318,6 +318,7 @@ func (p *pingfederateProvider) DataSources(_ context.Context) []func() datasourc
 		serversettingslogsettings.NewServerSettingsLogSettingsDataSource,
 		serversettingssystemkeys.NewServerSettingsSystemKeysDataSource,
 		sessionapplicationsessionpolicy.NewSessionApplicationSessionPolicyDataSource,
+		sessionauthenticationsessionpoliciesglobal.NewSessionAuthenticationSessionPoliciesGlobalDataSource,
 		virtualhostnames.NewVirtualHostNamesDataSource,
 	}
 }